Perf Testing on Rust node #181

+++ saito-core/src/core/data/blockchain.rs
@@ -1,6 +1,6 @@
+use rayon::prelude::*;
 use std::collections::VecDeque;
 use std::io::Error;
-
 use std::sync::Arc;
 
 use ahash::AHashMap;
@@ -21,7 +21,7 @@ use crate::core::mining_thread::MiningEvent;
 use crate::{log_write_lock_receive, log_write_lock_request};
 
 // length of 1 genesis period
-pub const GENESIS_PERIOD: u64 = 50;
+pub const GENESIS_PERIOD: u64 = 100;
 // prune blocks from index after N blocks
 pub const PRUNE_AFTER_BLOCKS: u64 = 10;
 // max recursion when paying stakers -- number of blocks including  -- number of blocks including GTT
@@ -59,7 +59,7 @@ impl Blockchain {
     #[allow(clippy::new_without_default)]
     pub fn new(wallet_lock: Arc<RwLock<Wallet>>) -> Self {
         Blockchain {
-            utxoset: AHashMap::new(),
+            utxoset: AHashMap::with_capacity(10_000_000),
             blockring: BlockRing::new(),
             blocks: AHashMap::new(),
             wallet_lock,
@@ -162,8 +162,10 @@ impl Blockchain {
                                     // log_read_lock_request!("mempool");
                                     // let mempool = mempool.read().await;
                                     // log_read_lock_receive!("mempool");
-                                    block_in_mempool_queue =
-                                        mempool.blocks_queue.iter().any(|b| block_hash == b.hash);
+                                    block_in_mempool_queue = mempool
+                                        .blocks_queue
+                                        .par_iter()
+                                        .any(|b| block_hash == b.hash);
                                 }
                                 if !block_in_mempool_queue {
                                     let result = network
@@ -1290,8 +1292,9 @@ impl Blockchain {
             // in either case, we are OK to throw out everything below the
             // lowest_block_id that we have found. we use the purge_id to
             // handle purges.
-            //
-            self.delete_blocks(purge_bid, storage).await;
+            if purge_bid > 0 {
+                self.delete_blocks(purge_bid, storage).await;
+            }
         }
 
         //TODO: we already had in update_genesis_period() in self method - maybe no need to call here?
@@ -1448,6 +1451,7 @@ mod tests {
     use crate::common::test_manager::test::TestManager;
     use crate::core::data::blockchain::{bit_pack, bit_unpack, Blockchain};
     use crate::core::data::wallet::Wallet;
+
     // use tracing_subscriber;
     // use tracing_subscriber::filter::Directive;
     // use tracing_subscriber::layer::SubscriberExt;
